back/chat_bot/requirements.txt
datetime
numpy
pandas
plotly
sklearn-pandas
stocknews
streamlit
tensorflow
yfinance
crewai
duckduckgo-search
bs4
undetected-chromedriver
langchain
langchain-google-vertexai
websocket-client
pandas
yfinance
plotly
streamlit
textblob
requests
newspaper3k